🧑‍💻 Support import of lamindb model classes before connecting to an instance
Adds support for importing lamindb model classes before running `ln.connect(...)`
via lazy import proxies for all objects that are only available post connect.
This plays much nicer with modern Python code style, where all imports are on the
top of the file. After the user connects to an instance, instantiating the proxy
classes returns instances of the corresponding models and calling classmethods
on the proxy delegates to the model class.